backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
idna==3.7
numpy==1.26.4
pandas==2.2.1
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.1
six==1.16.0
sqlparse==0.4.4
stripe==9.7.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1

